misc(stripe): Refact webhook handling
