+++ app/jobs/payment_providers/stripe/handle_event_job.rb
@@ -9,7 +9,7 @@ module PaymentProviders
       retry_on BaseService::NotFoundFailure
 
       def perform(organization:, event:)
-        result = PaymentProviders::StripeService.new.handle_event(
+        result = PaymentProviders::Stripe::HandleEventService.call(
           organization:,
           event_json: event
         )
