+++ app/services/payment_providers/stripe_service.rb
@@ -2,11 +2,6 @@
 
 module PaymentProviders
   class StripeService < BaseService
-    PAYMENT_SERVICE_CLASS_MAP = {
-      "Invoice" => Invoices::Payments::StripeService,
-      "PaymentRequest" => PaymentRequests::Payments::StripeService
-    }.freeze
-
     def create_or_update(**args)
       payment_provider_result = PaymentProviders::FindService.call(
         organization_id: args[:organization_id],
@@ -84,71 +79,6 @@ module PaymentProviders
       result.service_failure!(code: 'webhook_error', message: 'Invalid signature')
     end
 
-    def handle_event(organization:, event_json:)
-      event = ::Stripe::Event.construct_from(JSON.parse(event_json))
-      unless PaymentProviders::StripeProvider::WEBHOOKS_EVENTS.include?(event.type)
-        Rails.logger.warn("Unexpected stripe event type: #{event.type}")
-        return result
-      end
-
-      case event.type
-      when 'setup_intent.succeeded'
-        PaymentProviders::Stripe::Webhooks::SetupIntentSucceededService.call(
-          organization_id: organization.id,
-          event_json:
-        ).raise_if_error!
-      when 'customer.updated'
-        PaymentProviders::Stripe::Webhooks::CustomerUpdatedService.call(
-          organization_id: organization.id,
-          event_json:
-        ).raise_if_error!
-      when 'charge.succeeded'
-        payment_service_klass(event)
-          .new.update_payment_status(
-            organization_id: organization.id,
-            provider_payment_id: event.data.object.payment_intent,
-            status: 'succeeded',
-            metadata: event.data.object.metadata.to_h.symbolize_keys
-          ).raise_if_error!
-      when 'charge.dispute.closed'
-        PaymentProviders::Stripe::Webhooks::ChargeDisputeClosedService.call(
-          organization_id: organization.id,
-          event_json:
-        ).raise_if_error!
-      when 'payment_intent.payment_failed', 'payment_intent.succeeded'
-        status = (event.type == 'payment_intent.succeeded') ? 'succeeded' : 'failed'
-        payment_service_klass(event)
-          .new.update_payment_status(
-            organization_id: organization.id,
-            provider_payment_id: event.data.object.id,
-            status:,
-            metadata: event.data.object.metadata.to_h.symbolize_keys
-          ).raise_if_error!
-      when 'payment_method.detached'
-        PaymentProviderCustomers::StripeService
-          .new
-          .delete_payment_method(
-            organization_id: organization.id,
-            stripe_customer_id: event.data.object.customer,
-            payment_method_id: event.data.object.id,
-            metadata: event.data.object.metadata.to_h.symbolize_keys
-          ).raise_if_error!
-      when 'charge.refund.updated'
-        CreditNotes::Refunds::StripeService
-          .new.update_status(
-            provider_refund_id: event.data.object.id,
-            status: event.data.object.status,
-            metadata: event.data.object.metadata.to_h.symbolize_keys
-          )
-      end
-    rescue BaseService::NotFoundFailure => e
-      # NOTE: Error with stripe sandbox should be ignord
-      raise if event.livemode
-
-      Rails.logger.warn("Stripe resource not found: #{e.message}. JSON: #{event_json}")
-      BaseService::Result.new # NOTE: Prevents error from being re-raised
-    end
-
     private
 
     def unregister_webhook(stripe_provider, api_key)
@@ -165,14 +95,6 @@ module PaymentProviders
       Rails.logger.error(e.backtrace.join("\n"))
 
       Sentry.capture_exception(e)
-    end
-
-    def payment_service_klass(event)
-      payable_type = event.data.object.metadata.to_h[:lago_payable_type] || "Invoice"
-
-      PAYMENT_SERVICE_CLASS_MAP.fetch(payable_type) do
-        raise NameError, "Invalid lago_payable_type: #{payable_type}"
-      end
     end
   end
 end
