+++ spec/jobs/payment_providers/stripe/handle_event_job_spec.rb
@@ -3,7 +3,6 @@
 require 'rails_helper'
 
 RSpec.describe PaymentProviders::Stripe::HandleEventJob, type: :job do
-  let(:stripe_service) { instance_double(PaymentProviders::StripeService) }
   let(:result) { BaseService::Result.new }
   let(:organization) { create(:organization) }
 
@@ -12,9 +11,8 @@ RSpec.describe PaymentProviders::Stripe::HandleEventJob, type: :job do
   end
 
   before do
-    allow(PaymentProviders::StripeService).to receive(:new)
-      .and_return(stripe_service)
-    allow(stripe_service).to receive(:handle_event)
+    allow(PaymentProviders::Stripe::HandleEventService)
+      .to receive(:call)
       .and_return(result)
   end
 
@@ -24,7 +22,6 @@ RSpec.describe PaymentProviders::Stripe::HandleEventJob, type: :job do
       event: stripe_event
     )
 
-    expect(PaymentProviders::StripeService).to have_received(:new)
-    expect(stripe_service).to have_received(:handle_event)
+    expect(PaymentProviders::Stripe::HandleEventService).to have_received(:call)
   end
 end
