+++ spec/services/payment_providers/stripe/webhooks/setup_intent_succeeded_service_spec.rb
@@ -3,7 +3,7 @@
 require 'rails_helper'
 
 RSpec.describe PaymentProviders::Stripe::Webhooks::SetupIntentSucceededService, type: :service do
-  subject(:webhook_service) { described_class.new(organization_id: organization.id, event_json:) }
+  subject(:webhook_service) { described_class.new(organization_id: organization.id, event:) }
 
   let(:organization) { create(:organization) }
   let(:customer) { create(:customer, organization:) }
